Refactor file management utilities and update saveGeneratedCode implementation
